# T2S model inference optimization