I rewrite the inference code for t2smodel.infer_panel, which can achieves about 1.2\~1.5 times faster inference speed. 

I tested the the token generation speed on a Windows PC and an Ubuntu server with a Geforce 3090 GPU. The token generation speed on Windows w/o `flash-attention` is up to 110\~120 tokens/s, and that on Ubuntu is up to170\~180 tokens/s. I think the main reason of generation speed difference is that SPDA in windows does not compiled with flash attention. By the way, the inference speed w/ flash attention on linux is about 220~230 tokens/s, and on Windows is about 140~150 tokens/s.

![Generation Example](https://github.com/RVC-Boss/GPT-SoVITS/assets/143197337/600432a0-f1aa-4e0d-a3aa-6204528384b1)

======Update 1=====
I change implementation with TorchScript, and backup the codes with flash attention in `flash-attn` branch.