@wehos 2024-03-04
Hi, great work! I've verified it on my local RTX4090 card and it boosts the throughput from 55it/s to 85it/s. 

Regarding your reported gap between windows and Linux servers, I have a relevant observation. On another RTX4090 card server, with **identical docker image (ubuntu) and gpu cards**, it achieves 140it/s. I realized that the inference speed might also be subject to other hardware factors, e.g., cpu and motherboard. Anyway, the improvement on the same server is significant. Thanks again for the great work!