@wehos 2024-03-05
> > Hi, great work! I've verified it on my local RTX4090 card and it boosts the throughput from 55it/s to 85it/s.
> > Regarding your reported gap between windows and Linux servers, I have a relevant observation. On another RTX4090 card server, with **identical docker image (ubuntu-cuda12.1-cudnn8) and gpu cards**, it achieves 140it/s. I realized that the inference speed might also be subject to other hardware factors, e.g., cpu and motherboard. Anyway, the improvement on the same server is significant. Thanks again for the great work!
> > ======Update 1=====
> > Although the throughput diplayed by `tqdm` has a large gap, the overall inference speed is consistent across servers. Probably the gap across platforms (and servers) is due to the timing error for CUDA asynchronous operations.
> > ======Update 2=====
> > With **the same docker image (ubuntu-cuda12.1-cudnn8)**, another RTX3090-TI server is x2 faster than RTX4090. Looks like the motherboard plays an important role here...
> 
> Which version of PyTorch and Flash Attention are you used in your docker image? In addition, as ts2 model samples the results, it usually costs different time with the same input.

Pytorch 2.2.1 with latest flash attention. The throughput is not highly dependent on the output length. 

I will continue on the benchmarking but so far RTX4090 cannot achieve higher throughput than RTX3090-TI in any cases (please refer to my reply above).