@WatchTower-Liu 2024-03-05
The GPT model used is so small that inference on high-performance Gpus is difficult to create performance bottlenecks. More of the performance bottleneck should be in the performance of the iteration itself. In this process, high-performance Gpus are difficult to fully load.