@baicai-1145 2024-03-05
> **嗨，干得好！我已经在本地 RTX4090 卡上验证了它，它在同一服务器上将吞吐量从 55it/s 提高到 85it/s。**
> 
> 关于您报告的 Windows 和 Linux 服务器之间的差距，我有一个相关的观察。在另一台RTX4090卡服务器上，具有**相同的docker镜像（ubuntu-cuda12.1-cudnn8）和gpu卡**，它达到了140it/s。我意识到推理速度还可能受到其他硬件因素的影响，例如CPU和主板。无论如何，在同一个服务器上的改进是显着的。再次感谢您所做的出色工作！
> 
> ======更新1=====
> 
> 虽然显示的吞吐量差距`tqdm`较大，但各服务器的整体推理速度是一致的。跨平台（和服务器）之间的差距可能是由于 CUDA 异步操作的计时错误造成的。
> 
> ======更新2=====
> 
> 使用**相同的 docker 映像 (ubuntu-cuda12.1-cudnn8)**，另一台 RTX3090-TI 服务器比 RTX4090 快 2 倍。看来CPU在这里扮演着重要的角色......
> 
> 不同硬件上相同系统和软件环境的基准测试结果（20 次运行）：
> 
> RTX4090 + X99 + Xeon® E5-2686：`80it/s` RTX4090 + ROG CROSSHAIR + AMD Ryzen 9 5900x：`195it/s` RTX4090 + ROME2d32gm + AMD EPYC 7662：`100it/s` RTX4090 + B650 LiveMixer + AMD Ryzen 9 7900X： RTX3090Ti + ROG Strix + AMD Ryzen 9 5900x：**`400it/s`**`230it/s`
> 
> 看来GPU并不是当前在线推理的主要瓶颈。~如果有人可以实现比`230it/s`RTX4090 更高的吞吐量，我有兴趣了解系统设置。~

CPU的频率会比较大的影响推理速度