@GoHomeToMacDonal 2024-03-05
> **Hi, great work! I've verified it on my local RTX4090 card and it boosts the throughput from 55it/s to 85it/s on the same server.**
> 
> Regarding your reported gap between windows and Linux servers, I have a relevant observation. On another RTX4090 card server, with **identical docker image (ubuntu-cuda12.1-cudnn8) and gpu cards**, it achieves 140it/s. I realized that the inference speed might also be subject to other hardware factors, e.g., cpu and motherboard. Anyway, the improvement on the same server is significant. Thanks again for the great work!
> 
> ======Update 1=====
> 
> Although the throughput diplayed by `tqdm` has a large gap, the overall inference speed is consistent across servers. Probably the gap across platforms (and servers) is due to the timing error for CUDA asynchronous operations.
> 
> ======Update 2=====
> 
> With **the same docker image (ubuntu-cuda12.1-cudnn8)**, another RTX3090-TI server is x2 faster than RTX4090. Looks like the CPU plays an important role here...
> 
> Benchmark results (20 runs) with the same system and software environment on different hardwares:
> 
> RTX4090 + X99 + Xeon® E5-2686: `80it/s` RTX4090 + ROG CROSSHAIR + AMD Ryzen 9 5900x: `195it/s` RTX4090 + ROME2d32gm + AMD EPYC 7662: `100it/s` RTX4090 + B650 LiveMixer + AMD Ryzen 9 7900X: **`400it/s`** RTX3090Ti + ROG Strix + AMD Ryzen 9 5900x: `230it/s`
> 
> It seems GPU is not the major bottleneck for current online inference. ~If anyone can achieve higher throughput than `230it/s` with RTX4090, I would be interested to know the system setting.~

Thanks for you comprehensive benchmark on different hardwares. I implemented part of the inference codes with `TorchScript` w/o `flash-attention` and pushed the them to a new branch [torchscript](https://github.com/GoHomeToMacDonal/GPT-SoVITS/tree/torchscript). Although it does not leverage the `flash-attetntion` library, it achieves `150~160 tokens/s` on **Windows 11** + i5-12600K + RTX 3090, and `220~230 tokens/s` on Linux + AMD Ryzen 9 5950x + RTX 3090. 

I think this improvement can signifiantly benefit Windows users. If possible, could you help to test this implementation on more platfroms, especially on **RTX4090 + Xeon® E5-2686**?

By the way, to my knowledge, `flash-attention` 2.2+ implemented *flash decoding* to improve the token generation speed, but SDPA does not.